requirements.txt
mediapipe
opencv-python
numpy
Pillow


